pyproject.toml
[project]
name = "open-data-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
sail = ["pysail==0.4.2", "pyspark-client==4.0.0"]
